refactor: create mpc-crypto-types crate
