+++ crates/contract-interface/src/lib.rs
@@ -7,32 +7,30 @@ pub mod types {
         VerifiedAttestation, VerifiedDstackAttestation,
     };
     pub use config::{Config, InitConfig};
-    pub use conversions::CryptoConversionError;
-    pub use crypto::{
-        Bls12381G1PublicKey, Bls12381G2PublicKey, Ed25519PublicKey, PublicKey, Secp256k1PublicKey,
-    };
     pub use foreign_chain::*;
     pub use participants::{ParticipantId, ParticipantInfo, Participants};
 
     pub use metrics::Metrics;
-    pub use primitives::{
-        AccountId, CkdAppId, DomainId, Ed25519Signature, K256AffinePoint, K256Scalar,
-        K256Signature, SignatureResponse, Tweak,
-    };
+    pub use primitives::{AccountId, CkdAppId, DomainId, Tweak};
     pub use sign::*;
     pub use state::{
         AddDomainsVotes, AttemptId, AuthenticatedAccountId, AuthenticatedParticipantId,
         DomainConfig, DomainPurpose, DomainRegistry, EpochId, InitializingContractState, KeyEvent,
         KeyEventId, KeyEventInstance, KeyForDomain, Keyset, ProtocolContractState,
-        PublicKeyExtended, ResharingContractState, RunningContractState, SignatureScheme,
-        Threshold, ThresholdParameters, ThresholdParametersVotes,
+        ResharingContractState, RunningContractState, SignatureScheme, Threshold,
+        ThresholdParameters, ThresholdParametersVotes,
     };
     pub use updates::{ProposedUpdates, UpdateHash};
 
+    // Re-export crypto types from mpc-crypto-types
+    pub use mpc_crypto_types::{
+        Bls12381G1PublicKey, Bls12381G2PublicKey, CryptoConversionError, Ed25519PublicKey,
+        Ed25519Signature, K256AffinePoint, K256Scalar, K256Signature, ParsePublicKeyError,
+        PublicKey, PublicKeyExtended, Secp256k1PublicKey, SignatureResponse,
+    };
+
     mod attestation;
     mod config;
-    mod conversions;
-    mod crypto;
     mod foreign_chain;
     mod metrics;
     mod participants;
@@ -43,15 +41,15 @@ pub mod types {
 }
 
 #[cfg(feature = "blstrs")]
-pub use blstrs;
+pub use mpc_crypto_types::blstrs;
 
 #[cfg(feature = "near")]
-pub use near_sdk;
+pub use mpc_crypto_types::near_sdk;
 
 #[cfg(feature = "k256")]
-pub use k256;
+pub use mpc_crypto_types::k256;
 
 #[cfg(feature = "ed25519-dalek")]
-pub use curve25519_dalek;
+pub use mpc_crypto_types::curve25519_dalek;
 #[cfg(feature = "ed25519-dalek")]
-pub use ed25519_dalek;
+pub use mpc_crypto_types::ed25519_dalek;
