+++ crates/crypto-types/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "mpc-crypto-types"
+version = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[features]
+abi = ["borsh/unstable__schema", "schemars"]
+blstrs = ["dep:blstrs"]
+ed25519-dalek = ["dep:curve25519-dalek", "dep:ed25519-dalek"]
+k256 = ["dep:k256"]
+near = ["dep:near-sdk"]
+
+[dependencies]
+blstrs = { workspace = true, optional = true }
+borsh = { workspace = true }
+bs58 = { workspace = true }
+curve25519-dalek = { workspace = true, optional = true }
+derive_more = { workspace = true }
+ed25519-dalek = { workspace = true, optional = true }
+k256 = { workspace = true, optional = true }
+near-sdk = { workspace = true, optional = true }
+serde = { workspace = true }
+serde_with = { workspace = true }
+thiserror = { workspace = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+schemars = { workspace = true, optional = true }
+
+[dev-dependencies]
+assert_matches = { workspace = true }
+borsh = { workspace = true }
+group = { workspace = true }
+near-sdk = { workspace = true }
+rand = { workspace = true }
+serde_json = { workspace = true }
+
+[lints]
+workspace = true
