+++ crates/crypto-types/src/primitives.rs
@@ -1,7 +1,7 @@
 use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize};
+use serde_with::{hex::Hex, serde_as};
 
-/// A Near AccountId
 #[derive(
     Debug,
     Clone,
@@ -19,9 +19,13 @@ use serde::{Deserialize, Serialize};
     all(feature = "abi", not(target_arch = "wasm32")),
     derive(schemars::JsonSchema, borsh::BorshSchema)
 )]
-pub struct AccountId(pub String);
+#[serde(tag = "scheme")]
+pub enum SignatureResponse {
+    Secp256k1(K256Signature),
+    Ed25519 { signature: Ed25519Signature },
+}
 
-/// AppId for CKD
+#[serde_as]
 #[derive(
     Debug,
     Clone,
@@ -34,16 +38,74 @@ pub struct AccountId(pub String);
     Deserialize,
     BorshSerialize,
     BorshDeserialize,
-    derive_more::Into,
     derive_more::From,
+    derive_more::Deref,
+)]
+#[cfg_attr(
+    all(feature = "abi", not(target_arch = "wasm32")),
+    derive(schemars::JsonSchema, borsh::BorshSchema)
+)]
+pub struct Ed25519Signature(
+    #[cfg_attr(
+            all(feature = "abi", not(target_arch = "wasm32")),
+            schemars(with = "Vec<u8>") // Schemars doesn't support arrays of size greater than 32.
+        )]
+    #[serde_as(as = "[_; 64]")]
+    [u8; 64],
+);
+
+#[serde_as]
+#[derive(
+    Debug,
+    Clone,
+    Eq,
+    PartialEq,
+    Ord,
+    PartialOrd,
+    Hash,
+    Serialize,
+    Deserialize,
+    BorshSerialize,
+    BorshDeserialize,
+)]
+#[cfg_attr(
+    all(feature = "abi", not(target_arch = "wasm32")),
+    derive(schemars::JsonSchema, borsh::BorshSchema)
+)]
+pub struct K256Signature {
+    pub big_r: K256AffinePoint,
+    pub s: K256Scalar,
+    pub recovery_id: u8,
+}
+
+#[serde_as]
+#[derive(
+    Debug,
+    Clone,
+    Eq,
+    PartialEq,
+    Ord,
+    PartialOrd,
+    Hash,
+    Serialize,
+    Deserialize,
+    BorshSerialize,
+    BorshDeserialize,
+    derive_more::From,
+    derive_more::Into,
     derive_more::AsRef,
 )]
 #[cfg_attr(
     all(feature = "abi", not(target_arch = "wasm32")),
     derive(schemars::JsonSchema, borsh::BorshSchema)
 )]
-pub struct CkdAppId(pub [u8; 32]);
+/// AffinePoint on the Secp256k1 curve
+pub struct K256AffinePoint {
+    #[serde_as(as = "Hex")]
+    pub affine_point: [u8; 33],
+}
 
+#[serde_as]
 #[derive(
     Debug,
     Clone,
@@ -56,35 +118,15 @@ pub struct CkdAppId(pub [u8; 32]);
     Deserialize,
     BorshSerialize,
     BorshDeserialize,
-    derive_more::Into,
     derive_more::From,
+    derive_more::Into,
     derive_more::AsRef,
 )]
 #[cfg_attr(
     all(feature = "abi", not(target_arch = "wasm32")),
     derive(schemars::JsonSchema, borsh::BorshSchema)
 )]
-pub struct Tweak(pub [u8; 32]);
-
-#[derive(
-    Debug,
-    Clone,
-    Copy,
-    Eq,
-    PartialEq,
-    Ord,
-    PartialOrd,
-    Hash,
-    Serialize,
-    Deserialize,
-    BorshSerialize,
-    BorshDeserialize,
-    derive_more::Into,
-    derive_more::From,
-    derive_more::AsRef,
-)]
-#[cfg_attr(
-    all(feature = "abi", not(target_arch = "wasm32")),
-    derive(schemars::JsonSchema, borsh::BorshSchema)
-)]
-pub struct DomainId(pub u64);
+pub struct K256Scalar {
+    #[serde_as(as = "Hex")]
+    pub scalar: [u8; 32],
+}
